requirements.txt
# requirements.txt (Main Django Dependencies)
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt
django-cors-headers==4.3.1
django-storages==1.14.2
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.4
django-import-export==3.3.1
drf-spectacular==0.26.5
django-celery-results
django-celery-beat

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# Celery
celery==5.3.4
kombu==5.3.4

# Image processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# Authentication
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7

# File handling
boto3==1.34.0
django-storages[boto3]==1.14.2

# Utilities
python-decouple==3.8
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Development
django-debug-toolbar==4.2.0
ipython==8.17.2